requirements.txt
python-telegram-bot==13.17
transformers==4.33.2
torch
